+++ backend2/src/main/java/com/cmoney/backend2/authorization/service/AuthorizationService.kt
@@ -1,11 +1,13 @@
 package com.cmoney.backend2.authorization.service
 
 import com.cmoney.backend2.authorization.service.api.getexpiredtime.ExpiredTime
+import com.cmoney.backend2.authorization.service.api.hasauth.Auth
 import com.cmoney.backend2.base.model.calladapter.RecordApi
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.Path
+import retrofit2.http.Url
 
 /**
  * AuthorizationService
@@ -25,4 +27,11 @@ interface AuthorizationService {
         @Path("type") type: String,
         @Path("subjectId") subjectId: Long
     ): Response<ExpiredTime>
+
+    @RecordApi
+    @GET
+    suspend fun hasAuth(
+        @Header("Authorization") authorization: String,
+        @Url url: String
+    ): Response<Auth>
 }