+++ backend2/src/main/java/com/cmoney/backend2/brokerdatatransmission/service/api/brokerstockdata/put/PutBrokerStockDataRequest.kt
@@ -1,0 +1,12 @@
+package com.cmoney.backend2.brokerdatatransmission.service.api.brokerstockdata.put
+
+import com.google.gson.annotations.SerializedName
+
+class PutBrokerStockDataRequest(
+    @SerializedName("CountryISOCode")
+    val code: Int,
+    @SerializedName("Guid")
+    val guid: String,
+    @SerializedName("BrokerData")
+    val brokerData: BrokerData
+)