+++ buildSrc/src/main/kotlin/ModuleConfig.kt
@@ -1,5 +1,5 @@
 object ModuleConfig {
-    const val PROJECT_VERSION = "5.3.1"
+    const val PROJECT_VERSION = "5.6.0"
     //SDK
     const val COMPILE_SDK = 30
     const val BUILD_TOOLS_VERSION = "30.0.3"